+++ module/actor/actor.mjs
@@ -1913,14 +1913,20 @@ export class HeroSystem6eActor extends Actor {
                     parseInt(this.system.characteristics?.end?.max) - parseInt(this.system.characteristics?.end?.value),
                 hap: this.system.hap?.value,
                 heroicIdentity: this.system.heroicIdentity ?? true,
-                charges: this.items
+                resources: this.items
                     .filter(
                         (item) =>
-                            item.system.charges?.CHARGES &&
-                            (item.system.charges.max !== item.system.charges.value ||
-                                item.system.charges.clipsMax !== item.system.charges.clips),
+                            item.system.chargeItemModifier &&
+                            (item.system._charges !== item.system.chargesMax ||
+                                item.system._clips !== item.system.clipsMax ||
+                                item.system.ablative > 0),
                     )
-                    .map((o) => ({ id: o.id, charges: foundry.utils.deepClone(o.system.charges._source) })),
+                    .map((o) => ({
+                        id: o.id,
+                        _charges: o.system._charges,
+                        _clips: o.system._clips,
+                        ablative: o.system.ablative,
+                    })),
 
                 was5e: this.is5e,
             };
@@ -2461,17 +2467,18 @@ export class HeroSystem6eActor extends Actor {
 
             // retainValuesOnUpload Charges
             uploadProgressBar.advance(`${this.name}: retainValuesOnUpload charges`, 0);
-            for (const chargeData of retainValuesOnUpload.charges) {
+            for (const resourceData of retainValuesOnUpload.resources) {
                 // Careful: the HDC ID is intially a string, but coerced to Number in dataModel thus ==
-                const item = this.items.find((i) => i.id === chargeData.id);
+                const item = this.items.find((i) => i.id === resourceData.id);
                 if (item) {
                     await item.update({
-                        "system.charges.value": Math.min(item.system.charges.value, chargeData.charges.value),
-                        "system.charges.clips": Math.min(item.system.charges.clips, chargeData.charges.clips),
+                        "system._charges": Math.min(item.system._charges, resourceData._charges),
+                        "system._clips": Math.min(item.system._clips, resourceData._clips),
+                        "system.ablative": Math.max(item.system.ablative, resourceData.ablative),
                     });
                 } else {
                     console.warn(
-                        `Unable to locate ${chargeData.NAME}/${chargeData.ALIAS} to consume charges after upload.`,
+                        `Unable to locate ${resourceData.NAME}/${resourceData.ALIAS} to consume charges after upload.`,
                     );
                 }
             }