+++ module/combat.mjs
@@ -811,7 +811,7 @@ export class HeroSystem6eCombat extends Combat {
                                 (o.XMLID === "COSTSEND" && o.OPTIONID === "ACTIVATE") ||
                                 o.XMLID === "COSTSENDONLYTOACTIVATE",
                         )) || // Does the power use END continuously?
-                        (item.system.charges.CHARGES && !item.system.charges.continuing)), // Does the power use charges but is not continuous (as that is tracked by an effect when made active)?
+                        (item.system.chargeModifier && !item.system.chargeModifier.CONTINUING)), // Does the power use charges but is not continuous (as that is tracked by an effect when made active)?
             )) {
                 const {
                     error,